requirements.txt
pytube
pillow
ttkbootstrap
pyinstaller